rename api module to applications
